chore: simplify overengineered patterns across stores, AI prompts, dialogs, theme CSS
Five targeted simplifications identified from an overengineering audit:

1. **Pub-sub event bus** — trim dead event types and payload interfaces (13 declared, 2 used). Keep mechanism (it handles real cascade-delete cases where stores have circular import constraints).

2. **Confirmation dialog stack** — DeleteConfirmationDialog and GameSessionConfirmationDialog now delegate to ConfirmationDialog via variant. Strip unused `size`/`tone` props from SimpleModal (documented in source as no-op) and 11 callers passing them.

3. **Theme CSS triplication** — extract identical spacing scale and `--radius-full` to new `_shared-tokens.css`. Remove redundant duplicate token blocks from dark-mode selectors (cascade from light).

4. **Prompt builder cascade** — merge 6 satellite files (`narrativeGenerator.prompt.{context,inventory,itemAcquisition,itemLoss,lore,tone}.ts`) into one `narrativeGenerator.prompt.ts`. Keep personalization separate (300 LOC, substantive).

5. **Lore store action fragmentation** — merge 10 action files (`loreStore.actions.*.ts` + types) into one `loreStore.actions.ts`. Helper modules with real logic (extraction, resolution, dedup, utils) left as-is.

Net: 23 files deleted, 3 created. All 2,139 tests pass.
