+++ src/components/forms/WorldAttributesForm.tsx
@@ -205,8 +205,6 @@ const WorldAttributesForm: React.FC<WorldAttributesFormProps> = ({
         isOpen={showCreateModal} 
         onClose={() => setShowCreateModal(false)}
         title="Create Attribute"
-        size="xl"
-        
       >
         <div>
           Create a new custom attribute for this world.
@@ -227,8 +225,6 @@ const WorldAttributesForm: React.FC<WorldAttributesFormProps> = ({
         isOpen={!!editingAttribute} 
         onClose={() => setEditingAttribute(null)}
         title="Edit Attribute"
-        size="xl"
-        
       >
         <div>
           Modify the details of this world attribute.
