--- src/lib/ai/narrativeGenerator.prompt.types.ts
@@ -1,5 +1,0 @@
-export interface NarrativeStaticContentCache {
-  itemAcquisitionInstructions?: string;
-  itemLossInstructions?: string;
-  toneSettings?: Map<string, string>;
-}
