+++ src/lib/theme/themes/_shared-tokens.css
@@ -1,0 +1,22 @@
+/* ═══════════════════════════════════════════════════════════════
+   Shared tokens — identical across all themes (DS1/DS2/DS3) and modes.
+   Per-theme overrides live in ds1.css, ds2.css, ds3.css.
+   ═══════════════════════════════════════════════════════════════ */
+
+:root {
+  /* Spacing scale (Tailwind-aligned) */
+  --space-0_5: 0.125rem;
+  --space-1:   0.25rem;
+  --space-1_5: 0.375rem;
+  --space-2:   0.5rem;
+  --space-2_5: 0.625rem;
+  --space-3:   0.75rem;
+  --space-3_5: 0.875rem;
+  --space-4:   1rem;
+  --space-5:   1.25rem;
+  --space-6:   1.5rem;
+  --space-8:   2rem;
+
+  /* Universal radius token */
+  --radius-full: 9999px;
+}
