+++ src/lib/theme/themes/ds1.css
@@ -130,23 +130,9 @@
   --color-danger: rgb(185 28 28);
   --color-danger-hover: rgb(153 27 27);
 
-  /* Spacing scale (Tailwind-aligned) */
-  --space-0_5: 0.125rem;
-  --space-1:   0.25rem;
-  --space-1_5: 0.375rem;
-  --space-2:   0.5rem;
-  --space-2_5: 0.625rem;
-  --space-3:   0.75rem;
-  --space-3_5: 0.875rem;
-  --space-4:   1rem;
-  --space-5:   1.25rem;
-  --space-6:   1.5rem;
-  --space-8:   2rem;
-
-  /* Radius tokens */
+  /* Radius tokens (per-theme; --radius-full lives in _shared-tokens.css) */
   --radius-sm: 2px;
   --radius-md: 4px;
-  --radius-full: 9999px;
 }
 
 /* DS1 Dark Mode */
@@ -259,24 +245,6 @@
   --shadow-drawer: -12px 0 24px rgb(0 0 0 / 40%);
   --color-danger: rgb(239 68 68);
   --color-danger-hover: rgb(220 38 38);
-
-  /* Spacing scale (Tailwind-aligned) */
-  --space-0_5: 0.125rem;
-  --space-1:   0.25rem;
-  --space-1_5: 0.375rem;
-  --space-2:   0.5rem;
-  --space-2_5: 0.625rem;
-  --space-3:   0.75rem;
-  --space-3_5: 0.875rem;
-  --space-4:   1rem;
-  --space-5:   1.25rem;
-  --space-6:   1.5rem;
-  --space-8:   2rem;
-
-  /* Radius tokens */
-  --radius-sm: 2px;
-  --radius-md: 4px;
-  --radius-full: 9999px;
 }
 
 /* DS1 Background Pattern - Mechanical drafting grid */
